+++ Sources/PulseUI/Helpers/LoggerSync-watchos.swift
@@ -2,7 +2,7 @@
 //
 // Copyright (c) 2020–2023 Alexander Grebenyuk (github.com/kean).
 
-#if os(watchOS) || os(iOS)
+#if os(watchOS)
 
 import CoreData
 import Combine
@@ -12,22 +12,21 @@ import SwiftUI
 
 final class LoggerSyncSession: ObservableObject {
     @Published fileprivate(set) var fileTransferStatus: FileTransferStatus = .initial
-    
+
     private let delegate: SessionDelegate
     fileprivate var directory: TemporaryDirectory?
-    
+
     static let shared = LoggerSyncSession()
-    
+
     init() {
         self.delegate = SessionDelegate()
         self.delegate.session = self
-        
+
         if WCSession.isSupported() {
             WCSession.default.delegate = delegate
             WCSession.default.activate()
         }
     }
-    
     func transfer(store: LoggerStore) {
         let directory = TemporaryDirectory()
         let date = makeCurrentDate()
@@ -35,57 +34,21 @@ final class LoggerSyncSession: ObservableObject {
         Task {
             _ = try? await store.export(to: storeURL)
             let session = WCSession.default.transferFile(storeURL, metadata: nil)
-            self.fileTransferStatus = .sending(session.progress)
-            self.directory = directory
+            DispatchQueue.main.async {
+                self.fileTransferStatus = .sending(session.progress)
+                self.directory = directory
+            }
         }
     }
 }
 
 private final class SessionDelegate: NSObject, WCSessionDelegate {
     unowned var session: LoggerSyncSession!
-    
-#if os(iOS)
-    func sessionDidBecomeInactive(_ session: WCSession) {}
-    
-    func sessionDidDeactivate(_ session: WCSession) {}
-#endif
-    
+
     func session(_ session: WCSession, activationDidCompleteWith activationState: WCSessionActivationState, error: Error?) {}
-    
-#if os(iOS)
-    func session(_ session: WCSession, didReceive file: WCSessionFile) {
-        DispatchQueue.main.async {
-            do {
-                let directory = TemporaryDirectory()
-                let storeURL = directory.url.appendingPathComponent(file.fileURL.lastPathComponent, isDirectory: false)
-                try FileManager.default.moveItem(at: file.fileURL, to: storeURL)
-                
-                runHapticFeedback(.success)
-                ToastView {
-                    HStack {
-                        Image(systemName: "applewatch.watchface")
-                        Text("Store received")
-                        Spacer().frame(width: 16)
-                        Button("Open", action: {
-                            guard let store = try? LoggerStore(storeURL: storeURL) else {
-                                return
-                            }
-                            let vc = UIViewController.present { _ in
-                                NavigationView {
-                                    ConsoleView(store: store)
-                                }
-                            }
-                            vc?.onDeinit(directory.remove)
-                        }).foregroundColor(Color.blue)
-                    }
-                }.show()
-            } catch {
-                runHapticFeedback(.error)
-            }
-        }
-    }
-#endif
-    
+
+    func session(_ session: WCSession, didReceive file: WCSessionFile) {}
+
     func session(_ session: WCSession, didFinish fileTransfer: WCSessionFileTransfer, error: Error?) {
         DispatchQueue.main.async {
             self.session.directory?.remove()
@@ -106,7 +69,7 @@ enum FileTransferStatus {
     case sending(Progress)
     case failure(Error)
     case success
-    
+
     var title: String {
         switch self {
         case .initial:
@@ -119,7 +82,7 @@ enum FileTransferStatus {
             return "Store Sent"
         }
     }
-    
+
     var isButtonDisabled: Bool {
         switch self {
         case .initial:
@@ -134,4 +97,5 @@ struct FileTransferError: Identifiable {
     let id = UUID()
     let message: String
 }
+
 #endif