+++ Sources/PulseUI/Helpers/UIKit+Extensions.swift
@@ -5,7 +5,6 @@
 #if os(iOS)
 
 import UIKit
-import SwiftUI
 
 extension UIView {
     func pinToSuperview(insets: UIEdgeInsets = .zero) {
@@ -16,29 +15,6 @@ extension UIView {
             leadingAnchor.constraint(equalTo: superview!.leadingAnchor, constant: insets.left),
             bottomAnchor.constraint(equalTo: superview!.bottomAnchor, constant: -insets.bottom)
         ])
-    }
-}
-
-extension UIViewController {
-    @discardableResult
-    static func present<ContentView: View>(_ closure: (_ dismiss: @escaping () -> Void) -> ContentView) -> UIViewController? {
-        present { UIHostingController(rootView: closure($0)) }
-    }
-
-    @discardableResult
-    static func present(_ closure: (_ dismiss: @escaping () -> Void) -> UIViewController) -> UIViewController? {
-        guard let keyWindow = UIApplication.shared.windows.filter({ $0.isKeyWindow }).first,
-              var topController = keyWindow.rootViewController else {
-            return nil
-        }
-        while let presentedViewController = topController.presentedViewController {
-            topController = presentedViewController
-        }
-        let vc = closure({ [weak topController] in
-            topController?.dismiss(animated: true, completion: nil)
-        })
-        topController.present(vc, animated: true, completion: nil)
-        return vc
     }
 }